requirements.txt
# Audio capture and playback
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Voice Activity Detection
onnxruntime>=1.16.0

# ASR - Offline Whisper
faster-whisper>=1.0.0
ctranslate2>=4.0.0

# Translation - Offline
argostranslate>=1.9.0

# TTS - Offline (Piper)
piper-tts>=1.2.0

# Utilities
torch>=2.0.0
torchaudio>=2.0.0
